fix(media): move defaults to general
